build graph, check for permissions
